Modified the codes for ascci2bin and bin2ascii to read longer ranges of m and n

+++ Format_converter_programs/ascii2bin-mpi.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int main(int argc, char* argv[]){
+long long main(int argc, char* argv[]){
 	long int i,j;
 	float **a;
 	char filename[256],bin_file[256];
